Now checks if an index already exists before attempting to create one

+++ camel/app/tools/bowtie2/bowtie2index.py
@@ -62,7 +62,8 @@ class Bowtie2Index(Bowtie2):
             self._refgenome_fasta = multifasta_file
         else:
             self._refgenome_fasta = os.path.join(self._folder, self._tool_inputs['FASTA_REF'][0].basename)
-            if self._refgenome_fasta != self._tool_inputs['FASTA_REF'][0].path:
+            if self._refgenome_fasta != self._tool_inputs['FASTA_REF'][0].path and not os.path.exists(
+                    self._refgenome_fasta):
                 os.symlink(self._tool_inputs['FASTA_REF'][0].path, self._refgenome_fasta)
 
     def __set_output(self):
